==> VI_data_collection/1_response_data_cleaning/0_confirm_responses_1.0.Rmd <==
---
title: "Checking Responses"
author: "Andrew M. Demetriou"
date: "2023-04-03"
output: html_document
---

#AIM

This was an initial check of the response data prior to participant compensation. 

```{r setup, include=FALSE}
library('tidyverse')    # code logistics
library('data.table')   # read in .csv with fread
library('here')         # file logistics
library('ggplot2')      # data visualization
library('cowplot')      # multiple plots per figure
```

```{r}
# path with the actual data
data_file_path <- here('VI_data_collection', "_data", "survey_response_data")

# create a list with all of the files
data_files <- list.files(data_file_path)

# read in qualtrics data file
responses_dt <- fread(here(data_file_path, "annotation_run_1_April 3, 2023_09.15.csv"))

# remove junk on first two rows
responses_dt <- responses_dt[3:.N]

# read in prolific data file
demogs_dt <- fread(here(data_file_path, "prolific_export_637357bf150585a4675a4b6b.csv"))
demogs_dt[, PROLIFIC_PID := `Participant id`]

# merge datasets: PROLIFIC DATA NOT ON GITHUB
all_dt <- responses_dt[demogs_dt, on=.(PROLIFIC_PID)]
complete_dt <- responses_dt[demogs_dt, on=.(PROLIFIC_PID), nomatch=0]

# get question wording:
questions <- colnames(fread(here(data_file_path, data_files[1]), skip=1))

rm(data_file_path, data_files, demogs_dt, responses_dt, complete_dt)
```

```{r}
# remove identifying columns
de_identified_responses_dt <- all_dt[, ! c("IPAddress", "RecipientLastName", 
                         "RecipientFirstName", "RecipientEmail", 
                         "LocationLatitude", "LocationLongitude"), with = FALSE]

# select rows that are not test data
de_identified_responses_dt <- de_identified_responses_dt[DistributionChannel == "anonymous"]

rm(all_dt)
``` 

```{r}
# remove columns with unnecessary data
check_responses_dt <- de_identified_responses_dt[, ! c("StartDate", "EndDate", "Status", "RecordedDate", "ExternalReference", "DistributionChannel", "UserLanguage"), with=FALSE]

# check to see if these are useful columns
table(check_responses_dt$Finished)
table(check_responses_dt$`Participant Consent`)
table(check_responses_dt$Progress)

# remove columns
check_responses_dt <- check_responses_dt[, ! c("Progress", "Participant Consent", "Finished", "ResponseId"), with=FALSE]

rm(de_identified_responses_dt)
```

```{r}
check_responses_dt |>
  select(Prolific_ID, `Duration (in seconds)`) |>
  summarise(minutes = as.numeric(`Duration (in seconds)`) / 60) |>
  ggplot(aes(x=minutes)) +
  geom_histogram()
```
```{r}
#looking at the difference between Qualtrics time estimation and prolific

#Qualtrics
check_responses_dt$duration  <- as.numeric(check_responses_dt$`Duration (in seconds)`) 

#note columns to select
columns <- c("duration", "PROLIFIC_PID", "Time taken", "Sex", "Age")


duration_plot <- check_responses_dt |> select(all_of(columns)) |>
  ggplot(aes(x=duration/60, y=Age, color=Sex)) +
  geom_point()

#Prolific
time_taken_plot <- check_responses_dt %>% select(columns) %>%
  ggplot(., aes(x=`Time taken`/60, y=Age, color=Sex)) +
  geom_point()

plot_grid(duration_plot, time_taken_plot, ncol=1)
```


```{r}
long_response_times_dt <- check_responses_dt[duration/60 > 75,]
long_response_times_dt <- check_responses_dt[duration/60 > 80,]
long_response_times_dt <- check_responses_dt[duration/60 > 90,]

# these all look legit at first glance
rm(long_response_times_dt)
```

```{r}
# these responses left messages on the prolific platform

messages <- c("60f989f2b726f86c3d24e13e", # appears complete
              "6117efedb574017076d5b499", # appears complete
              "5e76c87c0fd39b22d52eba04") # wasn't able to participate

messages_dt <- check_responses_dt[check_responses_dt$PROLIFIC_PID %in% messages]
```
