==> VI_data_collection/1_response_data_cleaning/3_fix_NA_values_1.2.Rmd <==
---
title: "fix NA values"
author: "Andrew M. Demetriou"
date: "2023-04-06"
---

#AIM: Adjustment of NA values

Blank cells in the dataset can refer to multiple situations: e.g. data that was not entered vs. data that should be an explicit '0'. Specifically, participants were not shown every song lyric, and thus did not answer questions about the lyrics they were not shown. Further, participants had the option of indicating either a 0 or 'NA', which results in cells that show either blanks, or 0. 

This notebook adjusts those values such that:
1. NAs indicate cells for questions participants were NOT shown
2. 0s indicate cells for which participants indicated either 0 or NA to a value with reference to a specific song lyric. 



```{r}
library('here')
library('data.table')
library('dplyr')
library('stringr')
library('tidyr')

#source(here("VI_data_collection", '1_response_data_cleaning', '2_fix_column_names_1.1.R'))

source(here("VI_data_collection", '1_response_data_cleaning', '2_fix_column_names_1.2.R'))
```

```{r}
#make a temp dataframe
#ids <- all_dt[, ResponseId]
#temp_df <- all_dt[, 29:244]
#temp_df$PROLIFIC_PID <- ids
#temp_df$ResponseId <- ids
#temp_df <- temp_df |> select(PROLIFIC_PID, everything())
#temp_df <- temp_df |> select(ResponseId, everything())
#rm(ids)
```


```{r}
# character vector of relevant column names
relevant_columns <- c(  
  "-w", "-f", "1_1", "1_2", "1_3", "1_4", "1_5", 
  "1_6", "1_7", "1_8", "1_9", "1_10", "m_", "-c"
  )

cols <- all_dt |> select(contains(relevant_columns)) |> colnames()

# convert columns to numeric, blanks to NA
all_dt[,cols] <- lapply(cols, function(x)as.numeric(all_dt[[x]]))
complete_dt[,cols] <- lapply(cols, function(x)as.numeric(complete_dt[[x]]))

# convert free text column to NA
all_dt <- all_dt |> 
  #select(contains("-r")) |> 
  mutate(across(where(is.character), ~na_if(., "")))

complete_dt <- complete_dt |> 
  #select(contains("-r")) |> 
  mutate(across(where(is.character), ~na_if(., "")))

rm(cols, relevant_columns)
```

```{r}
# replace NAs with 0 when the participant has completed the question block:

# for every row
for(row in 1:nrow(all_dt)){
  
  # for every column
  for(col in 1:ncol(all_dt)){
    
    # if the column contains '-w'
    if(grepl("-w", colnames(all_dt[,..col]))==T){
      
      # if the column is not NA
      if(is.na(all_dt[row, ..col])==F){
       
        # for the next 11 columns in that row, replace them with:
        all_dt[row, seq(1,11)+col] <- setnafill(
          
          # the same values, but with the NA replaced with 0
          all_dt[row, seq(1,11)+col, with=FALSE], fill=0)
   }
  }
 }
}
```

```{r}
# replace NAs with 0 when the participant has completed the question block:

# for every row
for(row in 1:nrow(complete_dt)){
  
  # for every column
  for(col in 1:ncol(complete_dt)){
    
    # if the column contains '-w'
    if(grepl("-w", colnames(complete_dt[,..col]))==T){
      
      # if the column is not NA
      if(is.na(complete_dt[row, ..col])==F){
       
        # for the next 11 columns in that row, replace them with:
        complete_dt[row, seq(1,11)+col] <- setnafill(
          
          # the same values, but with the NA replaced with 0
          complete_dt[row, seq(1,11)+col, with=FALSE], fill=0)
   }
  }
 }
}
```


```{r}
# make missing data explicitly NA
for(i in names(all_dt[,PROLIFIC_PID:Finished])) (set(all_dt, which(all_dt[[i]] == "DATA_EXPIRED"), i, NA))
for(i in names(complete_dt[,PROLIFIC_PID:Finished])) (set(complete_dt, which(complete_dt[[i]] == "DATA_EXPIRED"), i, NA))
rm(i)

# replace missing values with NA
all_dt$Sex <- gsub("CONSENT_REVOKED", NA, all_dt$Sex)
complete_dt$Sex <- gsub("CONSENT_REVOKED", NA, complete_dt$Sex)

# not sure what this is meant to indicate
all_dt$Age <- gsub("1022", NA, all_dt$Age)
all_dt$Age <- gsub("CONSENT_REVOKED", NA, all_dt$Age)
all_dt$Age <- gsub("DATA_EXPIRED", NA, all_dt$Age)

complete_dt$Age <- gsub("1022", NA, complete_dt$Age)
complete_dt$Age <- gsub("CONSENT_REVOKED", NA, complete_dt$Age)
complete_dt$Age <- gsub("DATA_EXPIRED", NA, complete_dt$Age)

#t <- all_dt[is.na(all_dt$PROLIFIC_PID),]
```


```{r}
# intermediary working data
#saveRDS(all_dt, file = here("VI_data_collection", "_data","survey_response_data", "all_dt.RDS"))
#saveRDS(complete_dt, file = here("VI_data_collection", "_data","survey_response_data", "complete_dt.RDS"))
```


```{r}
# pivot raw data to long format to check
 all_dt_long <- all_dt |> 
  
  # pivot longer had an issue with mixed column types
  # make everything a character vector
  mutate(across(everything(), as.character)) |>
  
  # pivot to long format
  pivot_longer(
    
    # range of columns
    #cols = `828136-w`:`1027136-r`, 
    #cols = `828136-w`:`2107983-r`, 
    cols = `828136-w`:`213759518-r`, 
  
    # two new columns: song and question
    names_to = c('song', 'question'),
  
    # separate names by "-"
    names_sep = "-", 
  
    # new columns for participant responses
    values_to = "response", 
  
    # drop empty cells
    values_drop_na = TRUE
)

# pivot raw data to long format to check
 complete_dt_long <- complete_dt |> 
  
  # pivot longer had an issue with mixed column types
  # make everything a character vector
  mutate(across(everything(), as.character)) |>
  
  # pivot to long format
  pivot_longer(
    
    # range of columns
    #cols = `828136-w`:`1027136-r`, 
    #cols = `828136-w`:`2107983-r`, 
    cols = `828136-w`:`213759518-r`, 
  
    # two new columns: song and question
    names_to = c('song', 'question'),
  
    # separate names by "-"
    names_sep = "-", 
  
    # new columns for participant responses
    values_to = "response", 
  
    # drop empty cells
    values_drop_na = TRUE
)
```


```{r}
all_dt_long <- all_dt_long |> 
  pivot_wider(
    names_from = question, 
    values_from = response
  )

complete_dt_long <- complete_dt_long |> 
  pivot_wider(
    names_from = question, 
    values_from = response
  )

```

```{r}
s$participant_ID <- sapply(s$PROLIFIC_PID, digest::digest, algo = "md5") 

s<- s %>% select(-PROLIFIC_PID, -ResponseId, -hash, participant_ID, everything())
```
